[Next Gen] refactor(codegen): split fragment-shaped if-branch emission into branch_fragment
